Move phase poly generators to `generate`
